process_room/pyproject.toml
[project]
name = "process-room"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "pillow>=9.5.0",
]

[project.scripts]
process_room = "process_room:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"